+++ src/node/plan_node.cc
@@ -118,7 +118,8 @@ void PrintPlanVector(std::ostream &output,
     output << tab << SPACE_ST << vector_name << "[list]: \n";
     const std::string space = last_item ? (tab + INDENT) : tab + OR_INDENT;
     int i = 0;
-    for (i = 0; i < vec.size() - 1; ++i) {
+
+    for (i = 0; i < (int)(vec.size() - 1); ++i) {
         PrintPlanNode(output, space, vec[i], "" + std::to_string(i), false);
         output << "\n";
     }
